Decodes only what the reflect value is asking for.
As opposed to starting by iterating the text file. Another change coming
in this commit is a map where we are loading the VMX data, this was done
a way of reducing runtime complexity from O(n^2) to O(n) in exchange for
some more memory being used for the process.